fix after CR: add extractHybridQuery method to HybridQueryUtil
