+++ src/main/java/org/opensearch/neuralsearch/search/query/HybridCollectorManager.java
@@ -8,8 +8,6 @@ import java.util.Locale;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.lucene.index.IndexReader;
-import org.apache.lucene.search.BooleanClause;
-import org.apache.lucene.search.BooleanQuery;
 import org.apache.lucene.search.Collector;
 import org.apache.lucene.search.CollectorManager;
 import org.apache.lucene.search.FieldDoc;
@@ -61,9 +59,7 @@ import static org.opensearch.neuralsearch.search.util.HybridSearchResultFormatUt
 import static org.opensearch.neuralsearch.search.util.HybridSearchResultFormatUtil.createFieldDocStartStopElementForHybridSearchResults;
 import static org.opensearch.neuralsearch.search.util.HybridSearchResultFormatUtil.createFieldDocDelimiterElementForHybridSearchResults;
 import static org.opensearch.neuralsearch.search.util.HybridSearchResultFormatUtil.createSortFieldsForDelimiterResults;
-import static org.opensearch.neuralsearch.util.HybridQueryUtil.isHybridQueryExtendedWithDlsRulesAndWrappedInBoolQuery;
-import static org.opensearch.neuralsearch.util.HybridQueryUtil.isHybridQueryWrappedInBooleanQuery;
-import static org.opensearch.neuralsearch.util.HybridQueryUtil.isHybridQueryExtendedWithDlsRules;
+import static org.opensearch.neuralsearch.util.HybridQueryUtil.extractHybridQuery;
 
 /**
  * Collector manager based on HybridTopScoreDocCollector that allows users to parallelize counting the number of hits.
@@ -593,7 +589,7 @@ public abstract class HybridCollectorManager implements CollectorManager<Collect
      * @return results size to collected
      */
     private static int getSubqueryResultsRetrievalSize(final SearchContext searchContext) {
-        HybridQuery hybridQuery = unwrapHybridQuery(searchContext);
+        HybridQuery hybridQuery = extractHybridQuery(searchContext);
         Integer paginationDepth = hybridQuery.getQueryContext().getPaginationDepth();
 
         // Pagination is expected to work only when pagination_depth is provided in the search request.
@@ -607,46 +603,6 @@ public abstract class HybridCollectorManager implements CollectorManager<Collect
 
         // Switch to from+size retrieval size during standard hybrid query execution where from is 0.
         return searchContext.size();
-    }
-
-    /**
-     * Unwraps a HybridQuery from a direct query, a nested BooleanQuery, or a query extended with DLS rules by the security plugin.
-     */
-    private static HybridQuery unwrapHybridQuery(final SearchContext searchContext) {
-        HybridQuery hybridQuery;
-        Query query = searchContext.query();
-        if (isHybridQueryExtendedWithDlsRules(query, searchContext)) {
-            BooleanQuery booleanQuery = (BooleanQuery) query;
-            hybridQuery = unwrapHybridQueryWrappedInSecurityDlsRules(booleanQuery);
-        } else if (isHybridQueryExtendedWithDlsRulesAndWrappedInBoolQuery(searchContext, query)) {
-            BooleanQuery booleanQuery = (BooleanQuery) query;
-            hybridQuery = booleanQuery.clauses()
-                .stream()
-                .filter(clause -> isHybridQueryExtendedWithDlsRules(clause.query(), searchContext))
-                .findFirst()
-                .map(BooleanClause::query)
-                .map(BooleanQuery.class::cast)
-                .map(HybridCollectorManager::unwrapHybridQueryWrappedInSecurityDlsRules)
-                .orElseThrow(() -> new IllegalArgumentException("Given query does not contain a HybridQuery clause with DLS rules"));
-
-        } else if (isHybridQueryWrappedInBooleanQuery(searchContext, searchContext.query())) {
-            // In case of nested fields and alias filter, hybrid query is wrapped under bool query and lies in the first clause.
-            BooleanQuery booleanQuery = (BooleanQuery) query;
-            hybridQuery = (HybridQuery) booleanQuery.clauses().get(0).query();
-        } else {
-            hybridQuery = (HybridQuery) query;
-        }
-        return hybridQuery;
-    }
-
-    private static HybridQuery unwrapHybridQueryWrappedInSecurityDlsRules(BooleanQuery booleanQuery) {
-        return booleanQuery.clauses()
-            .stream()
-            .map(BooleanClause::query)
-            .filter(clauseQuery -> clauseQuery instanceof HybridQuery)
-            .map(HybridQuery.class::cast)
-            .findFirst()
-            .orElseThrow(() -> new IllegalArgumentException("Given boolean query does not contain a HybridQuery clause"));
     }
 
     /**
